format files better
